requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
PyMuPDF
pydantic
beautifulsoup4